Use brave theme type as a system dark mode on Windows
Aligning base UI theme(ex, context menu) with brave theme was not
implemented on Windows because Windows doesn't support per-application
theme. Whereas, MacOS support it. So, when user set dark or light, it is
set as a application theme.
To solve this on Windows, NativeThemeWin::SystemDarkModeEnabled() checks
whether dark mode enabling is overridden or not.
Overridden value is set by client(BraveThemeService).